esta_shit_esta_mal

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: alejogogi <alejogogi@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/29 09:47:50 by alejogogi         #+#    #+#             */
-/*   Updated: 2025/02/07 16:47:48 by alejogogi        ###   ########.fr       */
+/*   Updated: 2025/03/31 22:22:42 by alejogogi        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int		ft_putstr(char *s);
 int		ft_puthex(unsigned long long num, char c);
 int		ft_putunbr(unsigned int num);
 int		ft_putpointer(void *ptr);
-int		ft_putnbr(int num);
+int		ft_putnbr(long num);
 char	*ft_strchr(const char *s, int c);
 
 #endif
